+++ Lynx/Math/Vector3.cpp
@@ -48,7 +48,7 @@ Vector3::Vector3(const std::string& str) : x(0), y(0), z(0)
 {
 	//Removing spaces vector size if neccesary and splits into all the elements
 	std::vector<std::string> parts = strSplit(strSplit(str, ";")[0], " ");
-	for (unsigned int i = 0; i < min(3, parts.size()); i++)
+	for (size_t i = 0; i < min(3, parts.size()); i++)
 		(*this)[i] = numf(parts[i]);
 }
 
@@ -57,7 +57,7 @@ Vector3 Vector3::Parse(const std::string& str)
 	//Removing spaces vector size if neccesary and splits into all the elements
 	std::vector<std::string> parts = strSplit(strSplit(str, ";")[0], " ");
 	Vector3 result;
-	for (unsigned int i = 0; i < min(3, parts.size()); i++)
+	for (size_t i = 0; i < min(3, parts.size()); i++)
 		result[i] = numf(parts[i]);
 	return result;
 }
